requirements.txt

Flask
pillow
numpy
wiotp-sdk
scikit-image
keras
matplotlib
ibm-cos-sdk
ibm-cos-sdk-core
watson-machine-learning-client==1.0.371
tensorflow
sklearn


